New version format for development builds
This new version format for development builds follows the Semantic Versioning 2.0.0 format.

Instead of the previous method of time-based versioning, which can be tedious and confusing to manage, the previous final release will be denoted by a hyphen followed by a `alpha.x`.

We will begin at `v0.1.0-alpha.1` and continue from there.